+++ src/commands/database-instances-create.ts
@@ -3,8 +3,16 @@ import logger = require("../logger");
 import { requirePermissions } from "../requirePermissions";
 import { warnEmulatorNotSupported } from "../emulator/commandUtils";
 import { Emulators } from "../emulator/types";
-import { createInstance, DatabaseLocation, parseDatabaseLocation } from "../management/database";
+import {
+  createInstance,
+  DatabaseInstanceType,
+  DatabaseLocation,
+  parseDatabaseLocation,
+} from "../management/database";
 import getProjectId = require("../getProjectId");
+import { getDefaultDatabaseInstance } from "../getDefaultDatabaseInstance";
+import { FirebaseError } from "../error";
+import { MISSING_DEFAULT_INSTANCE_ERROR_MESSAGE } from "../requireDatabaseInstance";
 
 export default new Command("database:instances:create <instanceName>")
   .description("create a realtime database instance")
@@ -14,10 +22,20 @@ export default new Command("database:instances:create <instanceName>")
   )
   .before(requirePermissions, ["firebasedatabase.instances.create"])
   .before(warnEmulatorNotSupported, Emulators.DATABASE)
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   .action(async (instanceName: string, options: any) => {
     const projectId = getProjectId(options);
+    const defaultDatabaseInstance = await getDefaultDatabaseInstance({ project: projectId });
+    if (defaultDatabaseInstance === "") {
+      throw new FirebaseError(MISSING_DEFAULT_INSTANCE_ERROR_MESSAGE);
+    }
     const location = parseDatabaseLocation(options.location, DatabaseLocation.US_CENTRAL1);
-    const instance = await createInstance(projectId, instanceName, location);
+    const instance = await createInstance(
+      projectId,
+      instanceName,
+      location,
+      DatabaseInstanceType.USER_DATABASE
+    );
     logger.info(`created database instance ${instance.name}`);
     return instance;
   });