+++ src/commands/hosting-disable.ts
@@ -1,11 +1,12 @@
 import * as clc from "cli-color";
 
+import { Client } from "../apiv2";
 import { Command } from "../command";
+import { hostingApiOrigin } from "../api";
 import { promptOnce } from "../prompt";
-import { requirePermissions } from "../requirePermissions";
-import * as api from "../api";
-import * as utils from "../utils";
 import { requireHostingSite } from "../requireHostingSite";
+import { requirePermissions } from "../requirePermissions";
+import * as utils from "../utils";
 
 export default new Command("hosting:disable")
   .description("stop serving web traffic to your Firebase Hosting site")
@@ -31,13 +32,8 @@ export default new Command("hosting:disable")
       return;
     }
 
-    await api.request("POST", `/v1beta1/sites/${siteToDisable}/releases`, {
-      auth: true,
-      data: {
-        type: "SITE_DISABLE",
-      },
-      origin: api.hostingApiOrigin,
-    });
+    const c = new Client({ urlPrefix: hostingApiOrigin, apiVersion: "v1beta1", auth: true });
+    await c.post(`/sites/${siteToDisable}/releases`, { type: "SITE_DISABLE" });
 
     utils.logSuccess(
       `Hosting has been disabled for ${clc.bold(siteToDisable)}. Deploy a new version to re-enable.`