+++ src/emulator/auth/operations.ts
@@ -64,6 +64,7 @@ export const authOperations: AuthOps = {
         query: queryAccounts,
         sendOobCode,
         update: setAccountInfo,
+        batchGet,
       },
     },
   },
@@ -226,6 +227,33 @@ function lookup(
     // Drop users property if no users are found. This is needed for Node.js
     // Admin SDK: https://github.com/firebase/firebase-admin-node/issues/1078
     users: users.length ? users : undefined,
+  };
+}
+
+function batchGet(
+  state: ProjectState,
+  reqBody: unknown,
+  ctx: ExegesisContext
+): Schemas["GoogleCloudIdentitytoolkitV1DownloadAccountResponse"] {
+  const limit = Math.min(Math.floor(ctx.params.query.maxResults) || 20, 1000);
+  assert(limit >= 0, "((Auth Emulator: maxResults must not be negative.))");
+
+  const users = state.queryUsers(
+    {},
+    { sortByField: "localId", order: "ASC", startToken: ctx.params.query.nextPageToken }
+  );
+  let newPageToken: string | undefined = undefined;
+  if (users.length > limit) {
+    users.length = limit;
+    if (users.length) {
+      newPageToken = users[users.length - 1].localId;
+    }
+  }
+
+  return {
+    kind: "identitytoolkit#DownloadAccountResponse",
+    users,
+    nextPageToken: newPageToken,
   };
 }
 
@@ -948,7 +976,12 @@ function signInWithIdp(
   }
 
   let { response, rawId } = fakeFetchUserInfoFromIdp(providerId, claims);
-  response.oauthAccessToken = oauthAccessToken;
+
+  // Always return an access token, so that clients depending on it sorta work.
+  // e.g. JS SDK creates credentials from accessTokens for most providers:
+  // https://github.com/firebase/firebase-js-sdk/blob/6d640284ef6fd228bd7defdcb2d85a9f88239ad8/packages/auth/src/authcredential.js#L1515
+  response.oauthAccessToken =
+    oauthAccessToken || `FirebaseAuthEmulatorFakeAccessToken_${providerId}`;
   response.oauthIdToken = oauthIdToken;
   // What about response.refreshToken?
 
@@ -1278,7 +1311,7 @@ function parseIdToken(
   // TODO: Check JWT expiration here.
   const user = state.getUserByLocalId(decoded.payload.user_id);
   assert(user, "USER_NOT_FOUND");
-  assert(!user.validSince || decoded.payload.iat >= parseInt(user.validSince), "TOKEN_EXPIRED");
+  assert(!user.validSince || decoded.payload.iat >= Number(user.validSince), "TOKEN_EXPIRED");
   assert(!user.disabled, "USER_DISABLED");
 
   const signInProvider = decoded.payload.firebase.sign_in_provider;